Move UploadVODRequest to the only place it's used
