--- handlers/segmenting.go
@@ -1,16 +1,0 @@
-package handlers
-
-type UploadVODRequest struct {
-	Url             string `json:"url"`
-	CallbackUrl     string `json:"callback_url"`
-	OutputLocations []struct {
-		Type            string `json:"type"`
-		URL             string `json:"url"`
-		PinataAccessKey string `json:"pinata_access_key"`
-		Outputs         struct {
-			SourceMp4          bool `json:"source_mp4"`
-			SourceSegments     bool `json:"source_segments"`
-			TranscodedSegments bool `json:"transcoded_segments"`
-		} `json:"outputs,omitempty"`
-	} `json:"output_locations,omitempty"`
-}
